[Playground CLI] Move mounting code to mount.ts
